requirements.txt
Django==1.4.5
lxml==3.4.0
PyYAML==3.10
South==0.7.6
autopep8==0.8.6
distribute==0.6.30
django-admin-bootstrapped==0.3.2
django-countries==1.5
django-crispy-forms==1.2.3
django-guardian==1.2.4
django-transmeta==0.6.9
docutils==0.10
easy-thumbnails==1.2
pep8==1.4.1
pymongo==2.4.2
pyparsing==1.5.6
pytz==2012j
textile==2.1.5
wsgiref==0.1.2
djangorestframework==2.2.6
django-bootstrap-breadcrumbs==0.6.2
-e git://github.com/ribeiroct/django-bootstrap-pagination.git#egg=django-bootstrap-pagination
pysolr==3.0.6
github3.py==0.8.1
django-jenkins
-e git://github.com/bioinformatics-ua/django-userena.git#egg=userena
PIL==1.1.7
geopy==0.96.2
openpyxl==2.1.1
requests==2.4.3
#biopython==1.63
#django_nvd3==0.6.0
python-dateutil==2.2
celery==3.1.0
django-celery==3.1.0
python-memcached==1.48
-e git://github.com/jmoiron/johnny-cache.git@d96ea94c5dfcde517ff8f65d6ba4e435d8a0168c#egg=johnny-cache
django-extensions
vobject
chardet
surlex>=0.2.0
sorl-thumbnail
django-newsletter==0.4.1
-e git://github.com/ribeiroct/django-fack.git#egg=django-fack
django-adminplus==0.3
-e git://github.com/leandrosouza/django-hitcount.git#egg=django-hitcount
psycopg2==2.5.4
biopython==1.64
django-compressor==1.4
rjsmin==1.0.9
locustio==0.7.2
pyquery==1.2.9
raven==5.1.1
querystring-parser=1.2.0
